+++ lib/braintree/credit_card_verification_gateway.rb
@@ -46,11 +46,24 @@ module Braintree
 
     def self._create_signature
       [
-         {:options => [:amount, :merchant_account_id, :account_type]},
-         {:credit_card => [
-           :cardholder_name, :cvv, :expiration_date, :expiration_month, :expiration_year,
-           :number, {:billing_address => AddressGateway._shared_signature}
-         ]}
+        {:credit_card => [
+          :cardholder_name, :cvv, :expiration_date, :expiration_month, :expiration_year,
+          :number, {:billing_address => AddressGateway._shared_signature}
+        ]},
+        :intended_transaction_source,
+        {:options => [:amount, :merchant_account_id, :account_type]},
+        :payment_method_nonce,
+        :three_d_secure_authentication_id,
+        {:three_d_secure_pass_thru => [
+            :eci_flag,
+            :cavv,
+            :xid,
+            :three_d_secure_version,
+            :authentication_response,
+            :directory_response,
+            :cavv_algorithm,
+            :ds_transaction_id,
+        ]},
       ]
     end
   end