+++ spec/integration/braintree/credit_card_spec.rb
@@ -576,66 +576,6 @@ describe Braintree::CreditCard do
       end
     end
 
-    context "venmo_sdk" do
-      describe "venmo_sdk_payment_method_code" do
-        it "can pass a venmo sdk payment method code" do
-          customer = Braintree::Customer.create!
-          result = Braintree::CreditCard.create(
-            :customer_id => customer.id,
-            :venmo_sdk_payment_method_code => Braintree::Test::VenmoSDK::VisaPaymentMethodCode,
-          )
-          result.success?.should == true
-          result.credit_card.venmo_sdk?.should == false
-          result.credit_card.bin.should == "400934"
-          result.credit_card.last_4.should == "1881"
-        end
-
-        it "success? returns false when given an invalid venmo sdk payment method code" do
-          customer = Braintree::Customer.create!
-          result = Braintree::CreditCard.create(
-            :customer_id => customer.id,
-            :venmo_sdk_payment_method_code => Braintree::Test::VenmoSDK::InvalidPaymentMethodCode,
-          )
-
-          result.success?.should == false
-          result.message.should == "Invalid VenmoSDK payment method code"
-          result.errors.first.code.should == "91727"
-        end
-      end
-
-      describe "venmo_sdk_session" do
-        it "venmo_sdk? returns true when given a valid session" do
-          customer = Braintree::Customer.create!
-          result = Braintree::CreditCard.create(
-            :customer_id => customer.id,
-            :number => Braintree::Test::CreditCardNumbers::Visa,
-            :expiration_date => "05/2009",
-            :cvv => "100",
-            :options => {
-              :venmo_sdk_session => Braintree::Test::VenmoSDK::Session
-            },
-          )
-          result.success?.should == true
-          result.credit_card.venmo_sdk?.should == false
-        end
-
-        it "venmo_sdk? returns false when given an invalid session" do
-          customer = Braintree::Customer.create!
-          result = Braintree::CreditCard.create(
-            :customer_id => customer.id,
-            :number => Braintree::Test::CreditCardNumbers::Visa,
-            :expiration_date => "05/2009",
-            :cvv => "100",
-            :options => {
-              :venmo_sdk_session => Braintree::Test::VenmoSDK::InvalidSession
-            },
-          )
-          result.success?.should == true
-          result.credit_card.venmo_sdk?.should == false
-        end
-      end
-    end
-
     context "client API" do
       it "adds credit card to an existing customer using a payment method nonce" do
         nonce = nonce_for_new_payment_method(