+++ CHANGELOG.md
@@ -1,5 +1,15 @@
 # Changelog
 
+## 4.11.0
+* Add `intended_transaction_source` to `CreditCardVerificationGateway`
+* Add `payment_method_nonce` to `CreditCardVerificationGateway`
+* Add `three_d_secure_authentication_id` to `CreditCardVerificationGateway`
+* Add `three_d_secure_pass_thru` to `CreditCardVerificationGateway`
+* Add `network_tokenization_attributes` parameter to `Transaction.sale`
+* Add `network_token_details` to Transaction.sale response
+* Add `is_network_tokenized` to `CreditCardDetails`
+* Add subscriptions to `SepaDirectDebitAccount`
+
 ## 4.10.0
 * Add `billing_address` attribute reader to `ApplePayCard`
 * Add `pre_dispute_program` to `Dispute` and `DisputeSearch`